fixing bugs in migration down functions

+++ database/migrations/2022_09_02_134524_make_game_sessions_points_gained_decimal.php
@@ -26,7 +26,7 @@ return new class extends Migration
     public function down()
     {
         Schema::table('game_sessions', function (Blueprint $table) {
-            $table->tinyInteger('points_gained')->change();
+            $table->integer('points_gained')->change();
         });
     }
 };
